#88 Add class BaseActivity. Change leakCanary lib version.
