+++ app/build.gradle.kts
@@ -1,7 +1,3 @@
-import com.android.build.gradle.internal.api.BaseVariantOutputImpl
-import org.jetbrains.kotlin.config.KotlinCompilerVersion
-import org.jlleitschuh.gradle.ktlint.reporter.ReporterType
-
 plugins {
     id("com.android.application")
     id("kotlin-android")
@@ -103,7 +99,7 @@ dependencies {
     val navigationVersion = "2.1.0-alpha04"
     val koinVersion = "2.0.0-rc-1"
     val ankoVersion = "0.10.8"
-    val leakCanaryVersion = "1.6.3"
+    val leakCanaryVersion = "2.0-alpha-2"
 
     // Core
     implementation(kotlin("stdlib-jdk7", KotlinCompilerVersion.VERSION))
